Refactor store + opbeat tracks only in prod
