--- web/src/helpers/redux.js
@@ -1,16 +1,0 @@
-// Utility to react to store changes
-module.exports.observeStore = (store, select, onChange) => {
-  let currentState;
-
-  function handleChange() {
-    const nextState = select(store.getState());
-    if (nextState !== currentState) {
-      currentState = nextState;
-      onChange(currentState);
-    }
-  }
-
-  const unsubscribe = store.subscribe(handleChange);
-  handleChange();
-  return unsubscribe;
-};
